pyproject.toml
[tool.poetry]
name = "climfinrisk"
version = "0.1.0"
description = "Physical Climate Risk Modeling Platform for Financial Assets"
authors = ["Dr. Ian Helfrich <ian@gatech.edu>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ihelfrich/ClimFinRIsk"
repository = "https://github.com/ihelfrich/ClimFinRIsk"
keywords = ["climate", "risk", "finance", "geospatial", "modeling"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21.0"
pandas = "^1.3.0"
scipy = "^1.7.0"
scikit-learn = "^1.0.0"
xarray = "^0.19.0"
rasterio = "^1.2.0"
geopandas = "^0.10.0"
pyproj = "^3.2.0"
shapely = "^1.8.0"
fiona = "^1.8.0"
tensorly = "^0.7.0"
matplotlib = "^3.4.0"
seaborn = "^0.11.0"
plotly = "^5.0.0"
folium = "^0.12.0"
requests = "^2.25.0"
cdsapi = "^0.5.0"
jupyter = "^1.0.0"
notebook = "^6.4.0"
ipywidgets = "^7.6.0"
statsmodels = "^0.12.0"
pymc = "^4.0.0"
pyyaml = "^5.4.0"
tqdm = "^4.60.0"
python-dotenv = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.0"
pytest-cov = "^2.12.0"
black = "^22.0.0"
flake8 = "^4.0.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
